android_app/requirements.txt
kivy
requests
python-socketio[client]
pyjwt
pyaudio
pygame
